requirements.txt
eth-ape
ape-solidity
pytest
ipykernel
matplotlib
pandas